requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
flask>=2.3.0
flask-cors>=4.0.0

